requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
ollama==0.6.1
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-pptx==1.0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
xlsxwriter==3.2.9
